refactor: NumberGenerator 추상화
