+++ src/main/java/racingcar/NumberGenerator.java
@@ -1,0 +1,6 @@
+package racingcar;
+
+@FunctionalInterface
+public interface NumberGenerator {
+    int generateNumber();
+}
